+++ src/client.rs
@@ -233,7 +233,7 @@ fn apply_replication(
     messages: &mut ClientMessages,
     buffered_mutations: &mut BufferedMutations,
 ) {
-    for mut message in messages.receive(ServerChannel::Updates) {
+    for mut message in messages.drain_received(ServerChannel::Updates) {
         if let Err(e) = apply_update_message(world, params, &mut message) {
             error!("unable to apply update message: {e}");
 
@@ -252,7 +252,7 @@ fn apply_replication(
     let mutations_count = messages.received_count(ServerChannel::Mutations);
     if mutations_count != 0 {
         let mut acks = Vec::with_capacity(MutateIndex::POSTCARD_MAX_SIZE * mutations_count);
-        for message in messages.receive(ServerChannel::Mutations) {
+        for message in messages.drain_received(ServerChannel::Mutations) {
             if let Err(e) = buffer_mutate_message(params, buffered_mutations, message, &mut acks) {
                 error!("unable to buffer mutate message: {e}");
             }